+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/Datastore.java
@@ -38,6 +38,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import javax.net.ssl.SSLHandshakeException;
 
 /**
  * Datastore represents a proxy for the remote server, hiding details of the RPC layer. It:
@@ -53,6 +54,16 @@ import java.util.Set;
  * involved in actually making changes or reading data, and is otherwise stateless.
  */
 public class Datastore {
+
+  /**
+   * Error message to surface when Firestore fails to establish an SSL connection. A failed SSL
+   * connection likely indicates that the developer needs to provide an updated OpenSSL stack as
+   * part of their app's dependencies.
+   */
+  static final String SSL_DEPENDENCY_ERROR_MESSAGE =
+      "Could not establish a secure connection to the Firestore backend. This is likely a problem "
+          + "with your app, rather than with Firestore itself. See https://bit.ly/2XFpdma for "
+          + "instructions on how to enable TLS on Android 4.x devices.";
 
   /** Set of lowercase, white-listed headers for logging purposes. */
   static final Set<String> WHITE_LISTED_HEADERS =
@@ -206,6 +217,17 @@ public class Datastore {
       default:
         throw new IllegalArgumentException("Unknown gRPC status code: " + status.getCode());
     }
+  }
+
+  /**
+   * Determine whether the given status maps to the error that GRPC-Java throws when an Android
+   * device is missing required SSL Ciphers.
+   *
+   * <p>This error is non-recoverable and must be addressed by the app developer.
+   */
+  public static boolean isSslHandshakeError(Status status) {
+    return status.getCode().equals(Status.Code.UNAVAILABLE)
+        && status.getCause() instanceof SSLHandshakeException;
   }
 
   /**